orchestrator/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
SQLAlchemy==2.0.25
PyMySQL==1.1.0
cryptography==42.0.0

# Authentication
PyJWT==2.8.0

# Scheduling
APScheduler==3.10.4

# Docker SDK for spawning workers
docker==7.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
